Rename S1HandoverRequest function.
In the ns-3 version used in this repo the S1-AP interface (control plane
interface) is implemented only virtually (it still is in the latest ns-3
version if I am not mistaken) and therefore it could not be used to
simulate the required control plane packets exchange.
Therefore the X2-AP interface (which is implemented fully) is used
instead to signal the destination eNodeB of the upcoming handover (and
data is passed through the X2-U interface until the moment that traffic
is routed to the destination S/PGW from the packet core).

For the above reason rename the S1HandoverRequest function to
HandoverRequestWithPgwRelocation.